puts: all core models under core.models package in their respective files 	...makes changes accordingly
renamed:    core/time_utils.py -> core/timeutils.py

starts: app webapi for web backend
	...adds in settings.py

ignores: tests.py
