+++ .gitignore
@@ -4,3 +4,4 @@
 db.sqlite3
 */migrations/*
 !*/migrations/__init__.py
+tests.py
