+++ escapeerrands/settings.py
@@ -17,7 +17,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'core'
+    'core',
+    'webapi',
 ]
 
 MIDDLEWARE = [
